# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/standards/sets": {
      "get": {
        "operationId": "StandardsController_sets_api/v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Standard sets visible to me: shared ones and my district's",
        "tags": [
          "Standards"
        ]
      },
      "post": {
        "operationId": "StandardsController_createSet_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Standards"
        ]
      }
    },
    "/api/v1/standards/sets/import": {
      "post": {
        "operationId": "StandardsController_importCase_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImportCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import a 1EdTech CASE package (JSON file in field \"file\")",
        "tags": [
          "Standards"
        ]
      }
    },
    "/api/v1/standards/sets/{id}/standards": {
      "post": {
        "operationId": "StandardsController_createStandard_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStandardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Standards"
        ]
      }
    },
    "/api/v1/standards/sets/{id}": {
      "delete": {
        "operationId": "StandardsController_removeSet_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Standards"
        ]
      }
    },
    "/api/v1/standards": {
      "get": {
        "operationId": "StandardsController_list_api/v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields; prefix with - for descending, e.g. -createdAt,lastName",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "setId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "gradeLevel",
            "required": false,
            "in": "query",
            "schema": {
              "example": "7",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search standards by set, grade level and text",
        "tags": [
          "Standards"
        ]
      }
    },
...
    "/api/v1/report-cards/generate": {
      "post": {
        "operationId": "ReportCardsController_generate_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportCardsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or refresh draft report cards for a grading period (teachers: one class; administrators: the school)",
        "tags": [
          "Report cards"
        ]
      }
    },
    "/api/v1/report-cards/publish": {
      "post": {
        "operationId": "ReportCardsController_publishAll_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReportCardsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish every draft for a grading period; students and parents are told",
        "tags": [
          "Report cards"
        ]
      }
    },
    "/api/v1/report-cards": {
      "get": {
        "operationId": "ReportCardsController_list_api/v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields; prefix with - for descending, e.g. -createdAt,lastName",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "gradingPeriodId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published"
              ]
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "report_card",
                "progress"
              ]
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report cards: staff see their classes or school; students and parents see published ones",
        "tags": [
          "Report cards"
        ]
      }
    },
    "/api/v1/report-cards/{id}": {
      "get": {
        "operationId": "ReportCardsController_get_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Report cards"
        ]
      }
    },
    "/api/v1/report-cards/{id}/pdf": {
      "get": {
        "operationId": "ReportCardsController_pdf_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Report cards"
        ]
      }
    },
    "/api/v1/report-cards/{id}/lines/{lineId}": {
      "patch": {
        "operationId": "ReportCardsController_comment_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Teacher comment on one class line (while the card is a draft)",
        "tags": [
          "Report cards"
        ]
      }
    },
    "/api/v1/report-cards/{id}/publish": {
      "post": {
        "operationId": "ReportCardsController_publish_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Report cards"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartSchool API",
    "description": "LMS API for SmartSchool. Conventions: docs/09-API-DESIGN.md. Errors use RFC 9457 problem details.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSetDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "TEKS-MATH"
          },
          "name": {
            "type": "string",
            "example": "Texas Essential Knowledge and Skills: Mathematics"
          },
          "subject": {
            "type": "string",
            "example": "Mathematics"
          },
          "jurisdiction": {
            "type": "string",
            "example": "Texas"
          },
          "sourceUri": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "description": "Omit (district roles) for a set shared by every school"
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "ImportCaseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "CCSS-MATH",
            "description": "Code for the new set; defaults to the document title"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateStandardDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "CCSS.MATH.CONTENT.7.EE.A.1"
          },
          "description": {
            "type": "string"
          },
          "gradeLevels": {
            "type": "string",
            "example": "7"
          },
          "parentCode": {
            "type": "string",
            "description": "Code of the parent standard in the same set"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
...
      "GenerateReportCardsDto": {
        "type": "object",
        "properties": {
          "gradingPeriodId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "report_card",
              "progress"
            ],
            "default": "report_card"
          },
          "classId": {
            "type": "string",
            "format": "uuid",
            "description": "Only this class (teachers must give one); administrators may omit it for the whole school"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "gradingPeriodId"
        ]
      },
      "PublishReportCardsDto": {
        "type": "object",
        "properties": {
          "gradingPeriodId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "report_card",
              "progress"
            ],
            "default": "report_card"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "gradingPeriodId"
        ]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "comment"
        ]
      }
    }
  }
}